pre-deploy

+++ Posts/routers/posts.py
@@ -1,13 +1,16 @@
-from fastapi import APIRouter, Response, status, Form, File, UploadFile, HTTPException
+from fastapi import APIRouter, Response, status, Form, File, UploadFile, HTTPException, Depends
 from ..database import database, Post, Like
 from ..schemas import PostResponseModel, PostRequestModel, LikeDetail, LikeResponse
-from ..storage import bucket
+from ..storage import upload_image
+from ..tokens import decode_jwt, oauth2_scheme, InvalidTokenError, ExpiredTokenError
 
 from typing import List
 
+import httpx
+
 POST_LIMIT = 50
 
-router = APIRouter(prefix='/posts')
+router = APIRouter(prefix='/api/v1/posts')
 
 @router.get("", response_model=List[PostResponseModel])
 async def get_all_posts(page: int = 1, limit: int = 10):
@@ -24,12 +27,19 @@ async def get_all_posts(page: int = 1, limit: int = 10):
     
 
 @router.post('', response_model=PostResponseModel)
-async def create_post(username: str = Form(...), text: str = Form(...), image: UploadFile = File(...)):
+async def create_post(text: str = Form(...), image: UploadFile = File(...), token: str = Depends(oauth2_scheme)):
     
     try:
-        blob = bucket.blob(image.filename)
-        blob.upload_from_file(image.file, content_type=image.content_type)
-        image_url = blob.public_url
+        username = decode_jwt(token)
+        
+    except (InvalidTokenError, ExpiredSignatureError) as e:
+        raise HTTPException(status_code=401, detail=str(e))
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+    
+    try:
+        image_url = upload_image(image)
         
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Error uploading image: {str(e)}")
@@ -52,30 +62,65 @@ async def get_post(id: str):
         
     return post
 
+
 @router.put('/{id}', response_model=PostResponseModel)
-async def update_post(id: int, post_req: PostRequestModel):
+async def update_post(id: int, text: str | None = Form(None), image: UploadFile | None = File(None), token: str = Depends(oauth2_scheme)):
+    if text is None and image is None:
+        raise HTTPException(status_code=400, detail="Nothing to modify")
+
+    try:
+        username = decode_jwt(token)
+    except (InvalidTokenError, ExpiredSignatureError) as e:
+        raise HTTPException(status_code=401, detail=str(e))
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+
     post = Post.select().where(Post.id == id).first()
-    
+
     if post is None:
-        raise HTTPException(404, 'Post not found')
-    
-    post.text = post_req.text
-    post.imageurl = "modificado"
-    
+        raise HTTPException(status_code=404, detail='Post not found')
+
+    if post.username != username: 
+        raise HTTPException(status_code=403, detail="Not authorized to update this post")
+
+    if image is not None:
+        try:
+            image_url = upload_image(image)
+            post.imageurl = image_url
+        except Exception as e:
+            raise HTTPException(status_code=500, detail=f"Error uploading image: {str(e)}")
+
+    if text is not None:
+        post.text = text
+
     post.save()
-    
+
     return post
+    
      
 @router.delete('/{id}', response_model=PostResponseModel) 
-async def delete_post(id: str):
+async def delete_post(id: str, token: str = Depends(oauth2_scheme)):
     post = Post.select().where(Post.id == id).first()
     
     if post is None:
         raise HTTPException(404, 'Post not found')
     
+    try:
+        username = decode_jwt(token)
+    except (InvalidTokenError, ExpiredSignatureError) as e:
+        raise HTTPException(status_code=401, detail=str(e))
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+    
+    if post.username != username:
+        raise HTTPException(401, "Not your post")    
+        
     post.delete_instance()
     
     return post
+
 
 @router.get('/user/{username}', response_model=List[PostResponseModel])
 async def get_posts_of_user(username: str, page: int = 1, limit: int = 10):
@@ -93,16 +138,31 @@ async def get_posts_of_user(username: str, page: int = 1, limit: int = 10):
 
 
 @router.get("/following/all", response_model=List[PostResponseModel])
-async def get_posts_following(page: int = 1, limit: int = 10):
+async def get_posts_following(page: int = 1, limit: int = 10, token: str = Depends(oauth2_scheme)):
 
     if page <= 0 or limit <= 0:
         raise HTTPException(400, 'Params must be greater than zero')
     
     if limit > POST_LIMIT:
         raise HTTPException(400, 'Retreive less posts at the same time')
+    
+    try:
+        username = decode_jwt(token)
+        
+    except (InvalidTokenError, ExpiredSignatureError) as e:
+        raise HTTPException(status_code=401, detail=str(e))
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
 
-    usernames = ['Manolo', 'Roman', 'Valentina']
+    
+    
+    following_users = await get_following_users(username)
 
+    if not following_users:
+        raise HTTPException(404, "You must follow at least one user to search posts")
+    
+    usernames = [item['username'] for item in following_users]
+    
     posts = Post.select().where(Post.username.in_(usernames)).order_by(Post.created_at.desc()).paginate(page, limit)
     
     return [post for post in posts]
@@ -126,13 +186,23 @@ async def get_likes(id: str):
 
 
 @router.post('/{id}/like') 
-async def post_like(id: str):
+async def post_like(id: str, token: str = Depends(oauth2_scheme)):
+    try:
+        username = decode_jwt(token)
+        
+    except (InvalidTokenError, ExpiredSignatureError) as e:
+        raise HTTPException(status_code=401, detail=str(e))
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+    
+    
     post = Post.select().where(Post.id == id).first()
     
     if post is None:
         raise HTTPException(404, 'Post not found')
     
-    like, created = Like.get_or_create(post=post, username='Manolo')
+    like, created = Like.get_or_create(post=post, username=username)
     
     if not created:
         raise HTTPException(409, "Already liked")
@@ -141,13 +211,23 @@ async def post_like(id: str):
 
 
 @router.delete('/{id}/like') 
-async def delete_like(id: str):
+async def delete_like(id: str, token: str = Depends(oauth2_scheme)):
+    try:
+        username = decode_jwt(token)
+        
+    except (InvalidTokenError, ExpiredSignatureError) as e:
+        raise HTTPException(status_code=401, detail=str(e))
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+    
+    
     post = Post.select().where(Post.id == id).first()
     
     if post is None:
         raise HTTPException(404, 'Post not found')
     
-    like = Like.select().where((Like.post == post) & (Like.username == "Manolo")).first()
+    like = Like.select().where((Like.post == post) & (Like.username == username)).first()
     
     if like is None:
         raise HTTPException(404, 'Like not found')
@@ -155,3 +235,15 @@ async def delete_like(id: str):
     like.delete_instance()
     
     return Response("Disliked", status_code=status.HTTP_200_OK)
+
+
+
+async def get_following_users(username: str):
+    async with httpx.AsyncClient() as client:
+        url = f"http://127.0.0.1:8000/api/v1/profile/{username}/following"
+        response = await client.get(url)
+        if response.status_code != 200:
+            raise HTTPException(status_code=response.status_code, detail="Error fetching following users")
+        return response.json()
+    
+
